`:callback_url` option made required and `:region` - optional.
`lambda_process` is checking now if the Lambda function, specified in the uploader, is available, raising an error if not.

`lambda_process` is generating a random key, stores the key into the cached file metadata, and passes the key to the Lambda function for signing the request.

`lambda_authorize` method is fetching the signing key from the cached file metadata and use it for calculating the signature of the received from Lambda request - to compare the calculated and received signatures, returning an error if the signatures mismatch. If the signatures are equal, it returns the attacher and the hash of the parsed response from Lambda.

- Receives the `result` hash after Lambda request was authorized. The result could contain an array of processed file versions data hashes, or a single file data hash, if there were no versions and the original attached file was just moved to the target storage bucket.

'lambda_save` metog was changed to a Shrine Attacher instance method. It eletes the signing key, if it is present in the original file's metadata, converts the result to a JSON string, and writes this string into the `data_attribute` of the Shrine attacher's record.
Chooses the `save_methodz` either for the ActiveRecord or for Sequel, and saves the record.